Basic Vertices/Edges
Vertices and Edges split into 3 categories of complexity: 5 generics (complex), 2 generics (middle), and 1 generic (simple). Only simple has been documented thus far and there's a very good chance that complex doesn't work correctly.

I decided to split up to 5 and down to 1 for the following use cases:
SIMPLE (1) - Vertices simply connected with no edge weights or anything like that, just vertex content.
MIDDLE (2) - Vertices and edges each contain weight or content, so two generics are required. Additionally, edges can be "bundled" if they contain the same weight and the same set of "to" vertices (all vertices that fit this bill settle into the "from" vertices list).
COMPLEX (5) - Vertices and edges may be organized in layers, with each layer of edges and vertices potentially representing a different generic type (such as a Double edge pointing to a Double Vertex pointing to an Integer Edge pointing to a SpecialClass Vertex). This allows for more complex design of graphs, but also requires that each edge and vertex be moderately context-aware. See package-info for more information on the generics.